Camera request visuals done, doing back-end now
And so another journey has begun. After the fierce war of the passwordless login warriors agains the firebase servers everything went quiet and work resumed as normal. Now, a few days later a new war is threatening to start. Sages have foretold this war to be called the war of the permissions, according to legends this will either be a quick battle or a war that will be at large for days. Troups have already begun researching the enemy and how to win the battle so that when the time comes to fight we will be prepared. Now I will rest and gain some more energy for I shall be on the frontlines of this war and the probility of it already starting tomorrow is high.

~Goodnight